requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
geoalchemy2
psycopg2-binary
pydantic
pydantic_settings
python-dotenv
passlib[argon2]
python-jose[cryptography]
